Finished 4.21

+++ controllers/blogs.js
@@ -1,8 +1,10 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (req, res) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   res.json(blogs.map(blog => blog.toJSON()))
 })
 
@@ -16,16 +18,36 @@ blogsRouter.get('/:id', async (req, res, next) => {
   }
 })
 
+const getTokenFrom = req => {
+  const auth = req.get('authorization')
+  if (auth && auth.toLowerCase().startsWith('bearer ')) {
+    return auth.substring(7)
+  }
+  return null
+}
+
 blogsRouter.post('/', async (req, res, next) => {
   const body = req.body
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0
-  })
+  const token = getTokenFrom(req)
   try {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if(!token || !decodedToken.id) {
+      return res.status(401).json({ error: 'Token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+      user: user._id
+    })
+
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     res.json(savedBlog.toJSON())
   } catch (exception) {
     next(exception)
@@ -33,9 +55,19 @@ blogsRouter.post('/', async (req, res, next) => {
 })
 
 blogsRouter.delete('/:id', async (req, res, next) => {
+  const token = getTokenFrom(req)
   try {
-    await Blog.findByIdAndRemove(req.params.id)
-    res.status(204).end()
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if(!token || !decodedToken.id) {
+      return res.status(401).json({ error: 'Token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    if(user.blogs.includes(req.params.id)) {
+      await Blog.findByIdAndRemove(req.params.id)
+      res.status(204).end()
+    } else {
+      return res.status(401).json({ error: 'Token missing or invalid' })
+    }
   } catch (exception) {
     next(exception)
   }
